Move codebase to action package.
